Made config module
