+++ chain-signatures/contract/src/config/mod.rs
@@ -1,3 +1,7 @@
+mod impls;
+
+pub use impls::{min_to_ms, secs_to_ms};
+
 use std::collections::HashMap;
 
 use borsh::{self, BorshDeserialize, BorshSerialize};
@@ -11,46 +15,8 @@ pub struct Config {
     pub entries: HashMap<String, DynamicValue>,
 }
 
-impl Config {
-    pub fn get(&self, key: &str) -> Option<&serde_json::Value> {
-        let value = self.entries.get(key)?;
-        Some(&value.0)
-    }
-}
-
 #[derive(Clone, Debug, Serialize, Deserialize, PartialEq, Eq)]
 pub struct DynamicValue(serde_json::Value);
-
-impl From<serde_json::Value> for DynamicValue {
-    fn from(value: serde_json::Value) -> Self {
-        Self(value)
-    }
-}
-
-impl BorshSerialize for DynamicValue {
-    fn serialize<W: std::io::Write>(&self, writer: &mut W) -> std::io::Result<()> {
-        let buf = serde_json::to_vec(&self.0)
-            .map_err(|e| std::io::Error::new(std::io::ErrorKind::InvalidData, e))?;
-        BorshSerialize::serialize(&buf, writer)
-    }
-}
-
-impl BorshDeserialize for DynamicValue {
-    fn deserialize_reader<R: std::io::Read>(reader: &mut R) -> std::io::Result<Self> {
-        let buf: Vec<u8> = BorshDeserialize::deserialize_reader(reader)?;
-        let value = serde_json::from_slice(&buf)
-            .map_err(|e| std::io::Error::new(std::io::ErrorKind::InvalidData, e))?;
-        Ok(Self(value))
-    }
-}
-
-pub const fn secs_to_ms(secs: u64) -> u64 {
-    secs * 1000
-}
-
-pub const fn min_to_ms(min: u64) -> u64 {
-    min * 60 * 1000
-}
 
 #[cfg(test)]
 mod tests {
